+++ frontend/src/ui/forms/fixture-list-input.tsx
@@ -45,7 +45,7 @@ const scrollContainer = css`
 `
 
 const scrollContainer_compact = css`
-  height: ${baseline(64)};
+  height: ${baseline(48)};
 `
 
 const listEntry = css`